+++ phantom/utils/rllib/train.py
@@ -3,16 +3,7 @@ import tempfile
 from datetime import datetime
 from inspect import isclass
 from pathlib import Path
-from typing import (
-    Any,
-    Dict,
-    List,
-    Mapping,
-    Optional,
-    Tuple,
-    Type,
-    Union,
-)
+from typing import Any, Dict, List, Mapping, Optional, Tuple, Type, Union
 
 import cloudpickle
 import gymnasium as gym
@@ -26,11 +17,12 @@ from ray.rllib.evaluation import Episode, MultiAgentEpisode
 from ray.rllib.policy.sample_batch import SampleBatch
 from ray.rllib.utils.typing import TensorStructType, TensorType
 
-from ...agents import Agent
+from ...agents import Agent, StrategicAgent
 from ...env import PhantomEnv
 from ...metrics import Metric, logging_helper
 from ...policy import Policy
 from ...types import AgentID
+from ... import telemetry
 from .. import check_env_config, rich_progress, show_pythonhashseed_warning
 from .wrapper import RLlibEnvWrapper
 
@@ -76,7 +68,8 @@ def train(
         env_class: A PhantomEnv subclass.
         policies: A mapping of policy IDs to policy configurations.
         iterations: Number of training iterations to perform.
-        checkpoint_freq: The iteration frequency to save policy checkpoints at.
+        checkpoint_freq: The iteration frequency to save policy checkpoints at (defaults
+            to taking one checkpoint at the end of training).
         num_workers: Number of Ray rollout workers to use (defaults to 'NUM CPU - 1').
         env_config: Configuration parameters to pass to the environment init method.
         rllib_config: Optional algorithm parameters dictionary to pass to RLlib.
@@ -123,16 +116,21 @@ def train(
     if num_workers is not None:
         assert num_workers >= 0, "'num_workers' parameter must be >= 0"
 
+    if checkpoint_freq is None:
+        checkpoint_freq = iterations
+
     env_config = env_config or {}
     rllib_config = rllib_config or {}
     metrics = metrics or {}
 
     check_env_config(env_config)
 
-    ray.init(ignore_reinit_error=True, **(ray_config or {}))
-
     env = env_class(**env_config)
-    env.reset()
+    with telemetry.logger.pause():
+        env.validate()
+        env.reset()
+
+    ray.init(ignore_reinit_error=True, **(ray_config or {}))
 
     policy_specs: Dict[str, rllib.policy.policy.PolicySpec] = {}
     policy_mapping: Dict[AgentID, str] = {}
@@ -165,10 +163,13 @@ def train(
         else:
             raise TypeError(type(params))
 
+        agent = env.agents[agent_ids[0]]
+        assert isinstance(agent, StrategicAgent)
+
         policy_specs[policy_name] = rllib.policy.policy.PolicySpec(
             policy_class=policy_class,
-            action_space=env.agents[agent_ids[0]].action_space,
-            observation_space=env.agents[agent_ids[0]].observation_space,
+            action_space=agent.action_space,
+            observation_space=agent.observation_space,
             config=config,
         )
 
@@ -185,7 +186,7 @@ def train(
         env_class.__name__, lambda config: RLlibEnvWrapper(env_class(**config))
     )
 
-    num_workers_ = (os.cpu_count() - 1) if num_workers is None else num_workers
+    num_workers_ = ((os.cpu_count() or 1) - 1) if num_workers is None else num_workers
 
     timestr = datetime.today().strftime("%Y-%m-%d_%H-%M-%S")
     logdir_prefix = f"{algorithm}_{env.__class__.__name__}_{timestr}"
@@ -237,13 +238,13 @@ def train(
     algo = config_obj.build()
 
     with rich_progress("Training...") as progress:
-        for i in progress.track(range(iterations)):
+        for i in progress.track(range(1, iterations + 1)):
             result = algo.train()
 
             if show_training_metrics:
                 rich.pretty.pprint(
                     {
-                        "iteration": i + 1,
+                        "iteration": i,
                         "metrics": result["custom_metrics"],
                         "rewards": {
                             "policy_reward_min": result["policy_reward_min"],
@@ -253,7 +254,7 @@ def train(
                     }
                 )
 
-            if i == 0:
+            if i == 1:
                 config = {
                     "algorithm": algorithm,
                     "env_class": env_class,
@@ -270,7 +271,7 @@ def train(
                 with open(Path(algo.logdir, "phantom-training-params.pkl"), "wb") as f:
                     cloudpickle.dump(config, f)
 
-            if checkpoint_freq is not None and i % checkpoint_freq == 0:
+            if checkpoint_freq and (i % checkpoint_freq == 0 or i == iterations):
                 checkpoint_path = Path(algo.logdir, f"checkpoint_{str(i).zfill(6)}")
                 algo.save(checkpoint_path)
 
@@ -306,6 +307,8 @@ class RLlibMetricLogger(DefaultCallbacks):
 
 
 def make_rllib_wrapped_policy_class(policy_class: Type[Policy]) -> Type[rllib.Policy]:
+    """Internal function"""
+
     class RLlibPolicyWrapper(rllib.Policy):
         # NOTE:
         # If the action space is larger than -1.0 < x < 1.0, RLlib will attempt to